requirements.txt
einops==0.4.1
h5py==3.6.0
imath==0.0.2
matplotlib==3.6.2
numpy==1.23.4
open3d==0.16.0
opencv_python==4.6.0.66
OpenEXR==1.3.9
Pillow==9.4.0
PyYAML==6.0
tqdm==4.64.0
